refactor: 디렉토리 구조 및 헤더 파일 선언 목록 수정
1. 기능 함수 인자 및 함수명 변화에 따라 chess_utils.h 파일 선언 목록 수정
2. main.c 함수 단일화
3. 헤더 파일 존재에 따라 중복 코드 제거
4. display.c 파일 추가 : 보드 초기화 및 출력 담당
5. main.c 파일로부터 display 관련 함수를 chess_display.c 파일로 분리
6. chess_display.c 파일 코드 일부 수정
