--- main_feedback.txt
@@ -1,30 +1,0 @@
-
-
-// 플레이어 턴 진행 전
-char* gameFlag = is_checkmate(board, king);
-switch(gameFlag) {
-    case "checkmate" : 게임 끝내고 결과 발표 함수로 넘어가기
-    case "check" : "플레이어한테 현재 체크 상황이라고 알리고 턴 진행"
-    bool is_check = true; // 턴이 끝날 때 false로 변경
-    case "normal" ;
-        char* drawFlag = is_stalemate(board, king);
-            if(strcmp(drawFlag, "stalemate") == 0) : 게임 끝내고 결과 발표
-            else : 그냥 턴 진행
-}
-
-// 플레이어가 자신의 기물을 선택한 이후
-1. 킹을 골랐고, 그 킹의 moveHistory값이 0인 경우
-bool complete_castling = castling(board, king, false)
-if(complete_castling) break;
-else 마저 이동 경로 선택 단계 진행
-
-2. 폰을 고른 경우
-bool complete_en_passant = en_passant(board, pawn)
-if(complete_castling) break;
-else 마저 이동 경로 선택 단계 진행
-
-// 플레이어의 기물 이동 이후
-사용한 기물이 폰이고, 현재 위치가 적 진영 끝인가?
-player1 기준 y좌표 0
-player2 기준 y좌표 BOARD_SZIE - 1
-promotion(board, pawn)
